requirements.txt
annotated-types==0.6.0
anyio==3.7.1
Authlib==1.2.1
bcrypt==4.0.1
beanie==1.23.6
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
lazy-model==0.2.0
mongoengine==0.27.0
motor==3.3.1
packaging==23.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
replicate==0.19.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
